ocr-flask-system-main/ocr-flask-system-main/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
flask-cors
Flask-Redis
Flask-Caching
gunicorn
pytesseract
clean-text
python-decouple
python-dotenv
pytest
PyJWT
pluggy
packaging
MarkupSafe
itsdangerous
iniconfig
importlib-metadata
exceptiongroup
colorama
click
blinker
pytz
requests
alembic
beautifulsoup4
gunicorn
unidecode
